@@ -28,7 +28,7 @@
         "php-http/curl-client": "A PSR Compatible HTTP Client Library"
     },
     "require": {
-        "php": "^8.0",
+        "php": "^8.1",
         "ext-json": "*",
         "json-mapper/json-mapper": "^2.10",
         "psr/http-client": "^1.0",
