@@ -17,35 +17,28 @@ class VultrClient
 {
 	private VultrClientHandler $client;
 
-	private const MAP = [
-		'account'          => Services\Account\AccountService::class,
-		'applications'     => Services\Applications\ApplicationService::class,
-		'backups'          => Services\Backups\BackupService::class,
-		'baremetal'        => Services\BareMetal\BareMetalService::class,
-		'billing'          => Services\Billing\BillingService::class,
-		'blockstorage'     => Services\BlockStorage\BlockStorageService::class,
-		'dns'              => Services\DNS\DNSService::class,
-		'firewall'         => Services\Firewall\FirewallService::class,
-		'instances'        => Services\Instances\InstanceService::class, // TODO
-		'iso'              => Services\ISO\ISOService::class,
-		'kubernetes'       => Services\Kubernetes\KubernetesService::class, // TODO, do load balancers, and block storage before this.
-		'loadbalancers'    => Services\LoadBalancers\LoadBalancerService::class,
-		'objectstorage'    => Services\ObjectStorage\ObjectStorageService::class,
-		'operating_system' => Services\OperatingSystems\OperatingSystemService::class,
-		'plans'            => Services\Plans\PlanService::class,
-		'reserved_ip'      => Services\ReservedIP\ReservedIPService::class,
-		'regions'          => Services\Regions\RegionService::class,
-		'snapshots'        => Services\Snapshots\SnapshotService::class,
-		'ssh_keys'         => Services\SSHKeys\SSHKeyService::class,
-		'startup_scripts'  => Services\StartupScripts\StartupScriptService::class,
-		'users'            => Services\Users\UserService::class,
-		'vpc'              => Services\VPC\VPCService::class,
-	];
-
-	/**
-	 * Optimization
-	 */
-	private $class_cache = [];
+    public readonly Services\Account\AccountService $account;
+    public readonly Services\Applications\ApplicationService $applications;
+    public readonly Services\Backups\BackupService $backups;
+    public readonly Services\BareMetal\BareMetalService $baremetal;
+    public readonly Services\Billing\BillingService $billing;
+    public readonly Services\BlockStorage\BlockStorageService $blockstorage;
+    public readonly Services\DNS\DNSService $dns;
+    public readonly Services\Firewall\FirewallService $firewall;
+    public readonly Services\Instances\InstanceService $instances;
+    public readonly Services\ISO\ISOService $iso;
+    public readonly Services\Kubernetes\KubernetesService $kubernetes;
+    public readonly Services\LoadBalancers\LoadBalancerService $loadbalancers;
+    public readonly Services\ObjectStorage\ObjectStorageService $objectstorage;
+    public readonly Services\OperatingSystems\OperatingSystemService $operating_system;
+    public readonly Services\Plans\PlanService $plans;
+    public readonly Services\ReservedIP\ReservedIPService $reserved_ip;
+    public readonly Services\Regions\RegionService $regions;
+    public readonly Services\Snapshots\SnapshotService $snapshots;
+    public readonly Services\SSHKeys\SSHKeyService $ssh_keys;
+    public readonly Services\StartupScripts\StartupScriptService $startup_scripts;
+    public readonly Services\Users\UserService $users;
+    public readonly Services\VPC\VPCService $vpc;
 
 	/**
 	 * @param $http - PSR18 ClientInterface - https://www.php-fig.org/psr/psr-18/
@@ -61,6 +54,9 @@ class VultrClient
 		?StreamFactoryInterface $stream = null
 	)
 	{
+        foreach((new \ReflectionClass(self::class))->getProperties(\ReflectionProperty::IS_PUBLIC) as $property) {
+            unset($this->{$property->name});
+        }
 		try
 		{
 			$this->setClientHandler($API_KEY, $http, $request, $response, $stream);
@@ -84,11 +80,15 @@ class VultrClient
 
 	public function __get(string $name)
 	{
-		$class = self::MAP[$name] ?? null;
+        $type = (new \ReflectionClass(self::class))->getProperty($name)?->getType();
+        if(!$type instanceof \ReflectionNamedType) {
+            return null;
+        }
+        $class = $type->getName();
 
 		if ($class !== null)
 		{
-			return $this->class_cache[$class] ?? ($this->class_cache[$class] = new $class($this, $this->client));
+			return $this->$name ??=  new $class($this, $this->client);
 		}
 
 		return null;
